Split and tune tests, add data source fields and tests
