+++ vcd/resource_vcd_vapp_vm.go
@@ -381,14 +381,14 @@ func vmSchemaFunc(vmType typeOfVm) map[string]*schema.Schema {
 						Optional:     true,
 						Type:         schema.TypeString,
 						ValidateFunc: validation.StringInSlice([]string{"POOL", "DHCP", "MANUAL", "NONE"}, false),
-						Description:  "IP address allocation mode. One of POOL, DHCP, MANUAL, NONE",
+						Description:  "Secondary (IPv6) IP address allocation mode. One of POOL, DHCP, MANUAL, NONE",
 					},
 					"secondary_ip": {
 						Computed:     true,
 						Optional:     true,
 						Type:         schema.TypeString,
 						ValidateFunc: checkEmptyOrSingleIP(), // Must accept empty string to ease using HCL interpolation
-						Description:  "IP of the VM. Settings depend on `ip_allocation_mode`. Omitted or empty for DHCP, POOL, NONE. Required for MANUAL",
+						Description:  "Secondary (IPv6) IP of the VM. Settings depend on `secondary_ip_allocation_mode`. Omitted or empty for DHCP, POOL, NONE. Required for MANUAL",
 					},
 				},
 			},
